wip: move internal types into src/types, improve index.d.ts
