+++ packages/vite-plugin-svelte/src/types/compile.d.ts
@@ -1,7 +1,7 @@
 // eslint-disable-next-line node/no-missing-import
 import type { Processed } from 'svelte/types/compiler/preprocess';
-import type { SvelteRequest } from './id-types.d';
-import type { ResolvedOptions } from './options-types.d';
+import type { SvelteRequest } from './id.d.ts';
+import type { ResolvedOptions } from './options.d.ts';
 export type CompileSvelte = (
 	svelteRequest: SvelteRequest,
 	code: string,
