+++ packages/vite-plugin-svelte/src/types/dependencies.d.ts
