+++ packages/vite-plugin-svelte/src/types/esbuild.d.ts
