+++ packages/vite-plugin-svelte/src/types/id.d.ts
