+++ packages/vite-plugin-svelte/src/types/sourcemaps.d.ts
