+++ packages/vite-plugin-svelte/src/types/vite-plugin-svelte-cache.d.ts
