+++ packages/vite-plugin-svelte/src/types/vite-plugin-svelte-stats.d.ts
