+++ packages/vite-plugin-svelte/src/utils/load-svelte-config.ts
@@ -28,8 +28,8 @@ const dynamicImportDefault = new Function(
 /**
  *
  * @param {import('vite').UserConfig | undefined} viteConfig
- * @param {Partial<import('./options-types.d').Options>} inlineOptions
- * @returns {Promise<Partial<import('./options-types.d').SvelteOptions> | undefined>}
+ * @param {Partial<import('../index.d.ts').Options>} inlineOptions
+ * @returns {Promise<Partial<import('../index.d.ts').SvelteOptions> | undefined>}
  */
 export async function loadSvelteConfig(viteConfig, inlineOptions) {
 	if (inlineOptions?.configFile === false) {
@@ -92,7 +92,7 @@ export async function loadSvelteConfig(viteConfig, inlineOptions) {
 /**
  *
  * @param {import('vite').UserConfig | undefined} viteConfig
- * @param {Partial<import('./options-types.d').Options> | undefined} inlineOptions
+ * @param {Partial<import('../index.d.ts').Options> | undefined} inlineOptions
  * @returns {string | undefined}
  */
 function findConfigToLoad(viteConfig, inlineOptions) {
