+++ packages/vite-plugin-svelte/src/utils/vite-plugin-svelte-stats.ts
@@ -2,7 +2,7 @@ import { log } from './log.js';
 import { performance } from 'perf_hooks';
 import { normalizePath } from 'vite';
 
-/** @type {import('./vite-plugin-svelte-stats-types.d').CollectionOptions} */
+/** @type {import('../types/vite-plugin-svelte-stats.d.ts').CollectionOptions} */
 const defaultCollectionOptions = {
 	// log after 500ms and more than one file processed
 	logInProgress: (c, now) => now - c.collectionStart > 500 && c.stats.length > 1,
@@ -22,7 +22,7 @@ function humanDuration(n) {
 
 /**
  *
- * @param {import('./vite-plugin-svelte-stats-types.d').PackageStats[]} pkgStats
+ * @param {import('../types/vite-plugin-svelte-stats.d.ts').PackageStats[]} pkgStats
  * @returns {string}
  */
 function formatPackageStats(pkgStats) {
@@ -68,7 +68,7 @@ export class VitePluginSvelteStats {
 	// package directory -> package name
 	/** @type {import('./vite-plugin-svelte-cache.js').VitePluginSvelteCache} */
 	#cache;
-	/** @type {import('./vite-plugin-svelte-stats-types.d').StatCollection[]} */
+	/** @type {import('../types/vite-plugin-svelte-stats.d.ts').StatCollection[]} */
 	#collections = [];
 	/**
 	 *
@@ -80,20 +80,20 @@ export class VitePluginSvelteStats {
 	/**
 	 *
 	 * @param {string} name
-	 * @param {Partial<import('./vite-plugin-svelte-stats-types.d').CollectionOptions>=} opts
-	 * @returns {import('./vite-plugin-svelte-stats-types.d').StatCollection}
+	 * @param {Partial<import('../types/vite-plugin-svelte-stats.d.ts').CollectionOptions>=} opts
+	 * @returns {import('../types/vite-plugin-svelte-stats.d.ts').StatCollection}
 	 */
 	startCollection(name, opts) {
 		const options = {
 			...defaultCollectionOptions,
 			...opts
 		};
-		/** @type {import('./vite-plugin-svelte-stats-types.d').Stat[]} */
+		/** @type {import('../types/vite-plugin-svelte-stats.d.ts').Stat[]} */
 		const stats = [];
 		const collectionStart = performance.now();
 		const _this = this;
 		let hasLoggedProgress = false;
-		/** @type {import('./vite-plugin-svelte-stats-types.d').StatCollection} */
+		/** @type {import('../types/vite-plugin-svelte-stats.d.ts').StatCollection} */
 		const collection = {
 			name,
 			options,
@@ -106,7 +106,7 @@ export class VitePluginSvelteStats {
 				}
 				file = normalizePath(file);
 				const start = performance.now();
-				/** @type {import('./vite-plugin-svelte-stats-types.d').Stat} */
+				/** @type {import('../types/vite-plugin-svelte-stats.d.ts').Stat} */
 				const stat = { file, start, end: start };
 				return () => {
 					const now = performance.now();
@@ -132,7 +132,7 @@ export class VitePluginSvelteStats {
 
 	/**
 	 *
-	 * @param {import('./vite-plugin-svelte-stats-types.d').StatCollection} collection
+	 * @param {import('../types/vite-plugin-svelte-stats.d.ts').StatCollection} collection
 	 */
 	async #finish(collection) {
 		try {
@@ -144,7 +144,7 @@ export class VitePluginSvelteStats {
 				await this.#aggregateStatsResult(collection);
 				log.debug(
 					`${collection.name} done.\n${formatPackageStats(
-						/** @type {import('./vite-plugin-svelte-stats-types.d').PackageStats[]}*/ (
+						/** @type {import('../types/vite-plugin-svelte-stats.d.ts').PackageStats[]}*/ (
 							collection.packageStats
 						)
 					)}`,
@@ -171,7 +171,7 @@ export class VitePluginSvelteStats {
 
 	/**
 	 *
-	 * @param {import('./vite-plugin-svelte-stats-types.d').StatCollection} collection
+	 * @param {import('../types/vite-plugin-svelte-stats.d.ts').StatCollection} collection
 	 */
 	async #aggregateStatsResult(collection) {
 		const stats = collection.stats;
@@ -180,7 +180,7 @@ export class VitePluginSvelteStats {
 		}
 
 		// group stats
-		/** @type {Record<string,import('./vite-plugin-svelte-stats-types.d').PackageStats>} */
+		/** @type {Record<string,import('../types/vite-plugin-svelte-stats.d.ts').PackageStats>} */
 		const grouped = {};
 		stats.forEach((stat) => {
 			const pkg = /** @type {string} */ (stat.pkg);
